+++ internal/database/entity/dlcdata.go
@@ -1,40 +1,85 @@
 package entity
 
 import (
+	"database/sql/driver"
+	"errors"
+	"fmt"
+	"strconv"
+	"strings"
 	"time"
 
 	"gorm.io/gorm"
+	"gorm.io/gorm/logger"
 )
 
-// DLCData represents the db model of the oracle data rvalue/signature relative an asset
-type DLCData struct {
-	Base
-	PublishedDate time.Time `gorm:"primary_key"`
-	AssetID       string    `gorm:"primary_key"`
-	Rvalue        string    `gorm:"unique;not null"`
-	Signature     string
-	Value         string
-	Asset         Asset `gorm:"foreignkey:AssetID" json:"-"`
+// EventData represents the db model of the oracle data rvalue/signature relative an asset
+type EventData struct {
+	Timestamp
+	PublishedDate         time.Time   `gorm:"primary_key"`
+	AssetID               string      `gorm:"primary_key"`
+	Nonces                StringArray `gorm:"not null"`
+	Signatures            StringArray
+	Values                StringArray
+	Asset                 Asset `gorm:"foreignkey:AssetID" json:"-"`
+	AnnouncementSignature string
+	Base                  int
+	Unit                  string
+	IsSigned              bool
+	Precision             int
 
 	// TODO should be stored somewhere secure
-	Kvalue string `gorm:"unique;not null" json:"-"`
+	Kvalues StringArray `gorm:"not null" json:"-"`
 }
 
-// IsSigned returns true if the Signature is set
-func (m *DLCData) IsSigned() bool {
-	return m.Signature != ""
+// GetEventID returns the event ID for the given eventData structure
+func (eventData *EventData) GetEventID() string {
+	return eventData.AssetID + strconv.FormatInt(eventData.PublishedDate.Unix(), 10)
 }
 
-// CreateDLCData will try to create a DLCData with a new Rvalue corresponding to an asset and publishDate
+// StringArray is an alias type for an array of strings
+type StringArray []string
+
+// Scan implements the Scanner interface for gorm custom types
+func (s *StringArray) Scan(value interface{}) error {
+	csv, ok := value.(string)
+	if !ok {
+		return fmt.Errorf("Failed to unmarshal string value: %v", value)
+	}
+
+	*s = strings.Split(csv, ",")
+
+	return nil
+}
+
+// Value implements the Valuer interface for gorm custom types
+func (s StringArray) Value() (driver.Value, error) {
+	if s == nil || len(s) == 0 {
+		return nil, nil
+	}
+	return strings.Join(s, ","), nil
+}
+
+// GormDataType implements the GormDataTypeInterface for gorm custom types
+func (StringArray) GormDataType() string {
+	return "text"
+}
+
+// HasSignature returns true if the Signature is set
+func (eventData *EventData) HasSignature() bool {
+	return len(eventData.Signatures) > 0
+}
+
+// CreateEventData will try to create a DLCData with a new Rvalue corresponding to an asset and publishDate
 // if already in db, it will return the value found with no error
-func CreateDLCData(db *gorm.DB, assetID string, publishDate time.Time, signingk string, rvalue string) (*DLCData, error) {
+func CreateEventData(db *gorm.DB, assetID string, publishDate time.Time, signingks []string, rvalues []string, base int) (*EventData, error) {
 	tx := db.Begin()
 
-	newDLCData := &DLCData{
+	newDLCData := &EventData{
 		PublishedDate: publishDate,
 		AssetID:       assetID,
-		Kvalue:        signingk,
-		Rvalue:        rvalue,
+		Kvalues:       signingks,
+		Nonces:        rvalues,
+		Base:          base,
 	}
 
 	tx = tx.Create(newDLCData)
@@ -49,9 +94,9 @@ func CreateDLCData(db *gorm.DB, assetID string, publishDate time.Time, signingk 
 
 // FindDLCDataPublishedNear will try to retrieve the oldest dlcData which has been published between nearTime and rangeD
 // limit
-func FindDLCDataPublishedNear(db *gorm.DB, assetID string, nearTime time.Time, rangeD time.Duration) (*DLCData, error) {
-	dlcData := &DLCData{}
-	filterCondition := &DLCData{
+func FindDLCDataPublishedNear(db *gorm.DB, assetID string, nearTime time.Time, rangeD time.Duration) (*EventData, error) {
+	dlcData := &EventData{}
+	filterCondition := &EventData{
 		AssetID: assetID,
 	}
 	req := db.Where(filterCondition)
@@ -68,9 +113,9 @@ func FindDLCDataPublishedNear(db *gorm.DB, assetID string, nearTime time.Time, r
 
 // FindDLCDataPublishedBefore will try to retrieve the most recent dlcData which has been published BEFORE a specific
 // date
-func FindDLCDataPublishedBefore(db *gorm.DB, assetID string, nearTime time.Time) (*DLCData, error) {
-	dlcData := &DLCData{}
-	filterCondition := &DLCData{
+func FindDLCDataPublishedBefore(db *gorm.DB, assetID string, nearTime time.Time) (*EventData, error) {
+	dlcData := &EventData{}
+	filterCondition := &EventData{
 		AssetID: assetID,
 	}
 	req := db.Where(filterCondition)
@@ -86,9 +131,9 @@ func FindDLCDataPublishedBefore(db *gorm.DB, assetID string, nearTime time.Time)
 
 // FindDLCDataPublishedAt will try to retrieve asset dlcData at specific publish date
 // from database
-func FindDLCDataPublishedAt(db *gorm.DB, assetID string, publishDate time.Time) (*DLCData, error) {
-	dlcData := &DLCData{}
-	filterCondition := &DLCData{
+func FindDLCDataPublishedAt(db *gorm.DB, assetID string, publishDate time.Time) (*EventData, error) {
+	dlcData := &EventData{}
+	filterCondition := &EventData{
 		AssetID:       assetID,
 		PublishedDate: publishDate,
 	}
@@ -101,21 +146,24 @@ func FindDLCDataPublishedAt(db *gorm.DB, assetID string, publishDate time.Time) 
 
 // UpdateDLCDataSignatureAndValue will try to update signature and value of the DLCData if it exists
 // and if the DLCdata is not already signed
-func UpdateDLCDataSignatureAndValue(db *gorm.DB, assetID string, publishDate time.Time, sig string, value string) (*DLCData, error) {
-	tx := db.Begin()
-	filterCondition := &DLCData{
+func UpdateDLCDataSignatureAndValue(db *gorm.DB, assetID string, publishDate time.Time, sigs []string, values []string) (*EventData, error) {
+	filterCondition := &EventData{
 		AssetID:       assetID,
 		PublishedDate: publishDate,
 	}
-	tx = tx.Model(filterCondition)
-	// ensure that the signature and value are empty, doesn't work in using filterCondition (ignored)
-	tx = tx.Where("signature = ?", "").Where("value = ?", "")
+	db.Logger.LogMode(logger.Info)
+	tx := db.Begin()
+	tx = tx.Where(filterCondition)
 
-	tx = tx.Updates(DLCData{Signature: sig, Value: value})
-	if err := tx.Error; err != nil {
+	var old EventData
+	tx.First(&old)
+
+	if old.Signatures != nil || old.Values != nil {
 		tx.Rollback()
-		return nil, err
+		return nil, errors.New("Already signed or assigned values")
 	}
+
+	tx = tx.Updates(EventData{Signatures: sigs, Values: values})
 
 	if tx.RowsAffected == 0 {
 		tx.Rollback()