+++ internal/database/entity/dlcdata_test.go
@@ -6,12 +6,12 @@ import (
 	"testing"
 	"time"
 
-	"gorm.io/gorm"
 	"github.com/stretchr/testify/assert"
+	"gorm.io/gorm"
 )
 
 func GetInitializedDB() *gorm.DB {
-	db := test.NewOrm(&entity.Asset{}, &entity.DLCData{}).GetDB()
+	db := test.NewOrm(&entity.Asset{}, &entity.EventData{}).GetDB()
 	db.Create(&entity.Asset{AssetID: "test"})
 	return db
 }
@@ -19,20 +19,21 @@ func GetInitializedDB() *gorm.DB {
 func Test_CreateDLCData_NotPresent_ReturnsCorrectValue(t *testing.T) {
 	// arrange
 	db := GetInitializedDB()
-	expected := &entity.DLCData{
+	expected := &entity.EventData{
 		PublishedDate: time.Now().UTC(),
 		AssetID:       "test",
-		Rvalue:        "rvalue",
-		Kvalue:        "kvalue",
+		Nonces:        []string{"rvalue"},
+		Kvalues:       []string{"kvalue"},
 	}
 
 	// act
-	actual, err := entity.CreateDLCData(
+	actual, err := entity.CreateEventData(
 		db,
 		expected.AssetID,
 		expected.PublishedDate,
-		expected.Kvalue,
-		expected.Rvalue)
+		expected.Kvalues,
+		expected.Nonces,
+		2)
 
 	// assert
 	assertSub := assert.New(t)
@@ -43,16 +44,16 @@ func Test_CreateDLCData_NotPresent_ReturnsCorrectValue(t *testing.T) {
 func Test_CreateDLCData_Present_ReturnsError(t *testing.T) {
 	db := GetInitializedDB()
 	now := time.Now().UTC()
-	inDB := &entity.DLCData{AssetID: "test", PublishedDate: now, Kvalue: "kvalue1", Rvalue: "rvalue2"}
+	inDB := &entity.EventData{AssetID: "test", PublishedDate: now, Kvalues: []string{"kvalue1"}, Nonces: []string{"rvalue2"}}
 	db.Create(inDB)
-	_, err := entity.CreateDLCData(db, inDB.AssetID, inDB.PublishedDate, inDB.Kvalue, inDB.Rvalue)
+	_, err := entity.CreateEventData(db, inDB.AssetID, inDB.PublishedDate, inDB.Kvalues, inDB.Nonces, 2)
 	assert.Error(t, err)
 }
 
 func Test_FindDLCDataPublishedNear_NotPresent_ReturnsRecordNotFoundError(t *testing.T) {
 	db := GetInitializedDB()
 	now := time.Now()
-	expected := &entity.DLCData{AssetID: "test", PublishedDate: now.Add(time.Hour), Kvalue: "", Rvalue: ""}
+	expected := &entity.EventData{AssetID: "test", PublishedDate: now.Add(time.Hour), Kvalues: []string{""}, Nonces: []string{""}}
 	db.Create(expected)
 	_, err := entity.FindDLCDataPublishedNear(db, expected.AssetID, now, 30*time.Minute)
 	assert.EqualError(t, err, gorm.ErrRecordNotFound.Error())
@@ -61,7 +62,7 @@ func Test_FindDLCDataPublishedNear_NotPresent_ReturnsRecordNotFoundError(t *test
 func Test_FindDLCDataPublishedNear_Present_ReturnsCorrectValue(t *testing.T) {
 	db := GetInitializedDB()
 	now := time.Now()
-	expected := &entity.DLCData{AssetID: "test", PublishedDate: now.Add(time.Hour), Kvalue: "", Rvalue: ""}
+	expected := &entity.EventData{AssetID: "test", PublishedDate: now.Add(time.Hour), Kvalues: []string{""}, Nonces: []string{""}}
 	db.Create(expected)
 	actual, err := entity.FindDLCDataPublishedNear(db, expected.AssetID, now, 2*time.Hour)
 	assert.NoError(t, err)
@@ -72,7 +73,7 @@ func Test_FindDLCDataPublishedNear_Present_ReturnsCorrectValue(t *testing.T) {
 func Test_FindDLCDataPublishedBefore_NotPresent_ReturnsRecordNotFoundError(t *testing.T) {
 	db := GetInitializedDB()
 	now := time.Now()
-	expected := &entity.DLCData{AssetID: "test", PublishedDate: now.Add(time.Hour), Kvalue: "", Rvalue: ""}
+	expected := &entity.EventData{AssetID: "test", PublishedDate: now.Add(time.Hour), Kvalues: []string{""}, Nonces: []string{""}}
 	db.Create(expected)
 	_, err := entity.FindDLCDataPublishedBefore(db, expected.AssetID, now)
 	assert.EqualError(t, err, gorm.ErrRecordNotFound.Error())
@@ -81,7 +82,7 @@ func Test_FindDLCDataPublishedBefore_NotPresent_ReturnsRecordNotFoundError(t *te
 func Test_FindDLCDataPublishedBefore_Present_ReturnsCorrectValue(t *testing.T) {
 	db := GetInitializedDB()
 	now := time.Now()
-	expected := &entity.DLCData{AssetID: "test", PublishedDate: now.Add(-1 * time.Hour), Kvalue: "", Rvalue: ""}
+	expected := &entity.EventData{AssetID: "test", PublishedDate: now.Add(-1 * time.Hour), Kvalues: []string{""}, Nonces: []string{""}}
 	db.Create(expected)
 	actual, err := entity.FindDLCDataPublishedBefore(db, expected.AssetID, now)
 	assert.NoError(t, err)
@@ -99,7 +100,7 @@ func Test_FindDLCDataPublishedAt_NotPresent_ReturnsRecordNotFoundError(t *testin
 func Test_FindDLCDataPublishedAt_Present_ReturnsCorrectValue(t *testing.T) {
 	db := GetInitializedDB()
 	now := time.Now()
-	expected := &entity.DLCData{AssetID: "test", PublishedDate: now, Kvalue: "", Rvalue: ""}
+	expected := &entity.EventData{AssetID: "test", PublishedDate: now, Kvalues: []string{""}, Nonces: []string{""}}
 	db.Create(expected)
 	actual, err := entity.FindDLCDataPublishedAt(db, expected.AssetID, now)
 	assert.NoError(t, err)
@@ -111,13 +112,14 @@ func Test_UpdateDLCDataSignatureAndValue_Present_ReturnsUpdated(t *testing.T) {
 	// arrange
 	db := GetInitializedDB()
 	now := time.Now().UTC()
-	expected := &entity.DLCData{
+	expected := &entity.EventData{
 		AssetID:       "test",
 		PublishedDate: now,
-		Kvalue:        "",
-		Rvalue:        "",
-		Signature:     "new Signature",
-		Value:         "new value"}
+		Kvalues:       []string{""},
+		Nonces:        []string{""},
+		Signatures:    nil,
+		Values:        nil,
+	}
 	db.Create(expected)
 
 	// act
@@ -125,7 +127,7 @@ func Test_UpdateDLCDataSignatureAndValue_Present_ReturnsUpdated(t *testing.T) {
 		db,
 		expected.AssetID,
 		expected.PublishedDate,
-		expected.Signature, expected.Value)
+		expected.Signatures, expected.Values)
 
 	// assert
 	assertSub := assert.New(t)
@@ -133,11 +135,11 @@ func Test_UpdateDLCDataSignatureAndValue_Present_ReturnsUpdated(t *testing.T) {
 	assertDLCDataEqual(assertSub, expected, actual)
 }
 
-func assertDLCDataEqual(assertSub *assert.Assertions, expected *entity.DLCData, actual *entity.DLCData) {
+func assertDLCDataEqual(assertSub *assert.Assertions, expected *entity.EventData, actual *entity.EventData) {
 	assertSub.Equal(expected.AssetID, actual.AssetID)
 	assertSub.Equal(expected.PublishedDate, actual.PublishedDate)
-	assertSub.Equal(expected.Kvalue, actual.Kvalue)
-	assertSub.Equal(expected.Rvalue, actual.Rvalue)
-	assertSub.Equal(expected.Signature, actual.Signature)
-	assertSub.Equal(expected.Value, actual.Value)
+	assertSub.Equal(expected.Kvalues, actual.Kvalues)
+	assertSub.Equal(expected.Nonces, actual.Nonces)
+	assertSub.Equal(expected.Signatures, actual.Signatures)
+	assertSub.Equal(expected.Values, actual.Values)
 }