+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Enable decomposition of numerical event outcomes into digits signed separately using different nonces.
+
 ## [0.0.4] - 2020-26-10
 
 ### Changed